+++ src/smash.cc
@@ -316,11 +316,10 @@ ScatterActionsFinder actions_finder_for_dump(Configuration &configuration) {
  * Checks if there are unused config values.
  */
 void check_for_unused_config_values(const Configuration &configuration) {
-  const std::string report = configuration.to_string();
-
-  if (report != "{}") {
+  if (!configuration.is_empty()) {
     throw std::runtime_error(
-        "The following configuration values were not used:\n" + report);
+        "The following configuration values were not used:\n" +
+        configuration.to_string());
   }
 }